requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0
aiosmtplib==2.0.0
httpx~=0.28.1
celery~=5.5.0
aiofiles~=24.1.0
Unidecode~=1.3.8
redis~=5.2.1
starlette~=0.27.0
websockets~=11.0.3